2nd add user roles

+++ app/controllers/books_controller.rb
@@ -1,27 +1,34 @@
 class BooksController < ApplicationController
+  before_action :authenticate_user!, except: [:all_books, :index, :show]
   before_action :set_category, except: [:all_books]
   before_action :set_book, only: [:show, :update, :destroy]
   rescue_from ActiveRecord::RecordNotFound, with: :record_not_found
+  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
   def all_books
-    @books = Book.all
+    authorize Book
+    @books = Book.filter_by_status(params[:status]).page(params[:page]).per(3)
     render json: {
       data: @books,
       status: :ok,
-      message: "All books fetched successfully"
+      message: "All books fetched successfully",
+      meta: pagination_meta(@books)
     }
   end
 
   def index
-    @books = @category.books.all
+    authorize Book
+    @books_category = @category.books.page(params[:page]).per(2)
     render json: {
-      data: @books,
+      data: @books_category,
       status: :ok,
-      message: "Books fetched successfully"
+      message: "Books fetched successfully",
+      meta: pagination_meta(@books_category)
     }
   end
 
   def show
+    authorize Book
     render json: {
       data: @book,
       status: :ok,
@@ -30,6 +37,7 @@ class BooksController < ApplicationController
   end
 
   def create
+    authorize Book
     @book = @category.books.build(book_params)
     if @book.save
       render json: {
@@ -43,6 +51,7 @@ class BooksController < ApplicationController
   end
 
   def update
+    authorize @book
     if @book.update(book_params)
       render json: {
         data: @book,
@@ -62,6 +71,7 @@ class BooksController < ApplicationController
   end
 
   def destroy
+    authorize @book
     if @book.destroy
       render json: {
         data: nil,
@@ -74,6 +84,16 @@ class BooksController < ApplicationController
   end
 
   private
+
+  def pagination_meta(paginated_records)
+    {
+      current_page: paginated_records.current_page,
+      next_page: paginated_records.next_page,
+      prev_page: paginated_records.prev_page,
+      total_pages: paginated_records.total_pages,
+      total_count: paginated_records.total_count
+    }
+  end
 
   def set_category
     @category = Category.find(params[:category_id])
@@ -101,5 +121,11 @@ class BooksController < ApplicationController
 
   def record_not_found(resource)
     render json: { error: "#{resource} not found" }, status: :not_found
+  end
+
+  def user_not_authorized
+    render json: {
+      status: { code: 403, message: "You are not authorized to perform this action." }
+    }, status: :forbidden
   end
 end
